requirements.txt

# Core Dependencies
boto3>=1.34.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
openai>=1.12.0
chromadb>=0.4.22
tiktoken>=0.5.2
pydantic>=2.5.0
requests>=2.31.0

# Slack Dependencies
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# Additional utilities
python-dotenv>=1.0.0
